[maximal_planar] move Phase3 to own file
